tiktok_helper/requirements.txt
requests~=2.31.0
Pillow~=10.0.0
selenium~=4.11.2